Fix 4 engine correctness bugs surfaced by the blog CRUD example
All four reproduce against the ported engine; the C amalgamation is correct,
and after each fix the Zig output matches C (and `zig build test` stays green).

1. where.zig allocateIndexInfo — used a checked @intCast where C does a
   truncating (u8)op cast. An FTS5 query carrying an auxiliary-function
   constraint (snippet/highlight in `WHERE ... MATCH`) leaves high bits in
   `op`, so @intCast panicked. Use @truncate to match C exactly.

2. vdbe.zig sqlite3VdbeExec prologue — zeroed the per-statement time cache
   `iCurrentTime` as c_int (4 bytes); the field is sqlite3_int64. After the
   first datetime('now') cached a full i64, later statements only cleared the
   low half, leaving stale high bits -> the cache never reset and the clock
   "drifted" (2nd+ call returned a corrupted timestamp). Zero all 8 bytes.

3. expr.zig TK_TRIGGER codegen — sqlite3TableColumnToStorage was declared
   `extern ... c_int` but returns i16 in C; the x86-64 ABI leaves the upper
   return-register bits undefined for a 16-bit return, so the -1 (rowid)
   result read as 65535. That made OP_Param p1 = 65539 instead of 3, so a
   trigger referencing NEW.rowid / NEW.<INTEGER PRIMARY KEY> indexed far past
   the frame's register array and segfaulted. Declare the return as i16.

4. fts5_main.zig fts5PoslistBlob — force-unwrapped pPoslist (`pPoslist.?`),
   but an empty poslist is legitimately null with nByte==0; C passes the null
   through to sqlite3Fts5BufferAppendBlob, which no-ops when n==0. Crashed on
   `ORDER BY rank` over an external-content table. Accept ?[*]const u8.

+++ src/vdbe.zig
@@ -4273,7 +4273,10 @@ pub export fn sqlite3VdbeExec(p: ?*Vdbe) callconv(.c) c_int {
         done = true;
     } else {
         wr(c_int, p, Vdbe_rc, SQLITE_OK);
-        wr(c_int, p, Vdbe_iCurrentTime, 0);
+        // iCurrentTime is sqlite3_int64 — zero all 8 bytes. Writing it as c_int
+        // left the high 4 bytes set after the first datetime('now') cached a
+        // full i64, so the per-statement time cache never reset (drifting clock).
+        wr(i64, p, Vdbe_iCurrentTime, 0);
         wr(c_int, db, sqlite3_busyHandler_nBusy, 0);
         if (isInterrupted(db)) {
             exitKind = .abort_interrupt;
